systeme_de_gestion_de_stock/requirements.txt
asgiref==3.3.1
Django==3.1.7
django-crispy-forms==1.11.1
django-filter==2.4.0
psycopg2==2.8.6
pytz==2021.1
sqlparse==0.4.1
